Fix force close bug when clearing cache when no thumbnails are present

+++ src/com/boardgamegeek/DataHelper.java
@@ -587,7 +587,7 @@ public class DataHelper {
 	public static int deleteThumbnails() {
 		int count = 0;
 		File folder = getThumbnailFolder();
-		if (folder != null) {
+		if (folder != null && folder.listFiles() != null) {
 			for (File file : folder.listFiles()) {
 				if (file.delete()) {
 					count++;
